+++ example/bert_question_answer/android/app/src/main/kotlin/com/example/bert_question_answer/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.bert_question_answer
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}