+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# [0.3.0] - Audio Support and Task Library
+* Added support for audio-cases.
+* Task Library for Text based applications.
+
+# [0.2.2] - Image Package Update
+* Update Image APIs.
+
+# [0.2.1] - TFLite Flutter dependency update
+
+* Updated to tflite_flutter: ^0.9.0
+
 # [0.2.0] - Migrate to null-safety
 * Stable null-safety
 * Depends on tflite_flutter: ^0.8.0